requirements.txt
aiohttp==3.6.2
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
beautifulsoup4==4.8.2
billiard==3.6.3.0
boto3==1.12.41
botocore==1.15.41
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
cssselect==1.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-countries==6.1
django-elasticsearch-dsl==7.1.1
django-extensions==2.2.8
django-heroku==0.3.1
django-model-utils==4.0.0
django-redis==4.11.0
django-redis-cache==2.1.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
elasticsearch==7.7.1
elasticsearch-dsl==7.2.0
fake-useragent==0.1.11
firebase-admin==4.3.0
flower==0.9.4
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
humanize==0.5.1
idna==2.8
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.5
joblib==0.15.1
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
multidict==4.7.4
nltk==3.5
numpy==1.18.2
oauthlib==3.1.0
packaging==20.3
pandas==1.0.3
parse==1.14.0
pep8==1.7.1
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.12.2
psycopg2==2.7.5
psycopg2-binary==2.8.4
py==1.8.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyee==6.0.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pyppeteer==0.0.25
pyquery==1.4.1
pytest==5.4.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.10.5
python-http-client==3.2.4
python3-openid==3.1.0
pytz==2019.3
redis==3.3.1
regex==2020.5.14
requests==2.22.0
requests-aws4auth==0.9
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
selenium==3.141.0
sendgrid==6.1.1
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
telepot==12.7
tornado==6.0.4
tqdm==4.42.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
w3lib==1.21.0
wcwidth==0.1.9
webdriver-manager==2.3.0
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.11.2
yarl==1.4.2
zipp==3.1.0
